Cooked GenerationTransaction.Input.SignatureScriptPart1. Implemented BlockTemplate.CoinBaseAux. Moved log output to bin\Debug\logs.
