+++ src/CoiniumServ/CoiniumServ.csproj
@@ -91,6 +91,7 @@
     <Compile Include="Common\Extensions\StringExtensions.cs" />
     <Compile Include="Common\Helpers\Arrays\ArrayHelpers.cs" />
     <Compile Include="Common\Helpers\IO\FileHelpers.cs" />
+    <Compile Include="Common\Helpers\Time\TimeHelpers.cs" />
     <Compile Include="Common\Platform\Framework.cs" />
     <Compile Include="Common\Platform\PlatformManager.cs" />
     <Compile Include="Common\Versions\VersionInfo.cs" />
@@ -107,7 +108,6 @@
     <Compile Include="Core\Mining\Events\MinerAuthenticationEventArgs.cs" />
     <Compile Include="Core\Mining\JobCounter.cs" />
     <Compile Include="Core\Crypto\MerkleTree.cs" />
-    <Compile Include="Core\Mining\Mining.cs" />
     <Compile Include="Core\Coin\Daemon\Responses\BlockTemplateTransaction.cs" />
     <Compile Include="Core\Server\Vanilla\VanillaMiner.cs" />
     <Compile Include="Core\Mining\MiningManager.cs" />
