+++ src/CoiniumServ/Core/Server/Stratum/StratumService.cs
@@ -18,6 +18,7 @@
 
 using System;
 using AustinHarris.JsonRpc;
+using Coinium.Common.Extensions;
 using Coinium.Core.RPC.Sockets;
 using Coinium.Core.Server.Stratum.Responses;
 using Serilog;
@@ -44,6 +45,11 @@ namespace Coinium.Core.Server.Stratum
         public ulong ExtraNonceCounter { get; private set; }
 
         /// <summary>
+        /// ExtraNonce placeholder to be used with coinbase transactions.
+        /// </summary>
+        public static byte[] ExtraNoncePlaceholder;
+
+        /// <summary>
         /// The number of bytes that the miner users for its ExtraNonce2 counter 
         /// <remarks>Represents expected length of extranonce2 which will be generated by the miner. (http://mining.bitcoin.cz/stratum-mining)</remarks>
         /// </summary>
@@ -51,6 +57,7 @@ namespace Coinium.Core.Server.Stratum
 
         public StratumService()
         {
+            ExtraNoncePlaceholder = "f000000ff111111f".ToByteArray();
             this.GenerateInstanceId(); // generate instance id for the service.
             this.InitExtraNonceCounter(); // init. the extra nonce counter.
         }
